Move validation code to new class
