+++ tools/tcg_rim_tool/src/main/java/hirs/swid/Main.java
@@ -12,13 +12,15 @@ public class Main {
         Commander commander = new Commander();
         JCommander jc = JCommander.newBuilder().addObject(commander).build();
         jc.parse(args);
-        SwidTagGateway gateway = new SwidTagGateway();
+        SwidTagGateway gateway;
+        SwidTagValidator validator;
 
         if (commander.isHelp()) {
             jc.usage();
             System.out.println(commander.printHelpExamples());
         } else {
             if (!commander.getVerifyFile().isEmpty()) {
+                validator = new SwidTagValidator();
                 System.out.println(commander.toString());
                 String verifyFile = commander.getVerifyFile();
                 String rimel = commander.getRimEventLog();
@@ -28,7 +30,7 @@ public class Main {
                         gateway.setRimEventLog(rimel);
                     }
                     try {
-                        gateway.validateSwidTag(verifyFile);
+                        validator.validateSwidTag(verifyFile);
                     } catch (IOException e) {
                         System.out.println("Error validating RIM file: " + e.getMessage());
                         System.exit(1);
@@ -38,6 +40,7 @@ public class Main {
                     System.exit(1);
                 }
             } else {
+                gateway = new SwidTagGateway();
                 System.out.println(commander.toString());
                 String createType = commander.getCreateType().toUpperCase();
                 String attributesFile = commander.getAttributesFile();
