+++ tools/tcg_rim_tool/src/main/java/hirs/swid/SwidTagConstants.java
@@ -69,6 +69,9 @@ public class SwidTagConstants {
     public static final String SUPPORT_RIM_FORMAT = "supportRIMFormat";
     public static final String SUPPORT_RIM_URI_GLOBAL = "supportRIMURIGlobal";
 
+    public static final QName _SHA256_HASH = new QName(
+            "http://www.w3.org/2001/04/xmlenc#sha256",
+            "hash", "SHA256");
     public static final QName _COLLOQUIAL_VERSION = new QName(
             "http://csrc.nist.gov/ns/swid/2015-extensions/1.0",
             "colloquialVersion", "n8060");
