+++ tools/tcg_rim_tool/src/test/java/hirs/swid/TestSwidTagGateway.java
@@ -16,6 +16,7 @@ import org.testng.annotations.Test;
 
 public class TestSwidTagGateway {
 	private SwidTagGateway gateway;
+	private SwidTagValidator validator;
 	private final String DEFAULT_OUTPUT = "generated_swidTag.swidtag";
 	private final String DEFAULT_WITH_CERT = "generated_with_cert.swidtag";
 	private final String DEFAULT_NO_CERT = "generated_no_cert.swidtag";
@@ -28,6 +29,7 @@ public class TestSwidTagGateway {
 	public void setUp() throws Exception {
 		gateway = new SwidTagGateway();
 		gateway.setRimEventLog(supportRimFile);
+		validator = new SwidTagValidator();
 	}
 
 	@AfterClass
@@ -70,7 +72,7 @@ public class TestSwidTagGateway {
 	@Test
 	public void testValidateSwidTag() {
 	    try {
-	        Assert.assertTrue(gateway.validateSwidTag(TestSwidTagGateway.class.getClassLoader().getResource(DEFAULT_WITH_CERT).getPath()));
+	        Assert.assertTrue(validator.validateSwidTag(TestSwidTagGateway.class.getClassLoader().getResource(DEFAULT_WITH_CERT).getPath()));
 	    } catch (IOException e) {
 	        Assert.fail("Invalid swidtag!");
 	    }
